+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 - Added `formality` script.
+- Added `jasper` script.
 
 ### Changed
 - Update dependencies, with slight restructuring updating `clap` to v3.1, bumps minimum rust to 1.54